+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/ui/statusbar/StatusBarFactory.kt
@@ -134,8 +134,8 @@ class WidgetBar(project: Project) : EditorBasedWidget(project), StatusBarWidget.
                 }
                 addSeparator("Updates")
                 addAll(
-                    state.availableUpdates.filter {
-                        it.value.data != null && it.key != state.foundPackageManager
+                    state.availableUpdates.filter { (packageName, currentState) ->
+                        currentState.data != null && packageName != state.foundPackageManager
                     }.toSortedMap()
                         .map { (key, _) ->
                             DumbAwareAction.create(key) {
@@ -168,8 +168,8 @@ class WidgetBar(project: Project) : EditorBasedWidget(project), StatusBarWidget.
             Status.SCANNING_FOR_DEPRECATIONS -> "Scanning for deprecations (${state.scannedDeprecations}/${state.totalPackages})..."
             Status.SCANNING_FOR_PACKAGE_MANAGER -> "Scanning for package manager updates..."
             Status.READY -> {
-                val outdated =
-                    state.availableUpdates.filter { it.value.data != null }.size + if (state.foundPackageManager != null) 1 else 0
+                val outdated = state.availableUpdates.filter { it.value.data != null }.size +
+                        if (state.foundPackageManager != null) 1 else 0
                 val deprecated = state.deprecations.filter { it.value.data != null }.size
                 when (NUDSettingsState.instance.statusBarMode) {
                     StatusBarMode.FULL -> when {