+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## [Unreleased]
 
+## [3.3.0] - 2024-10-07
+
 ### Added
 
 - Add warnings for unmaintained dependencies (#140)
@@ -305,7 +307,8 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.2.1...HEAD
+[Unreleased]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.3.0...HEAD
+[3.3.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.2.1...v3.3.0
 [3.2.1]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.0.2...v3.1.0